fastapi-app/requirements.txt
alembic==1.16.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
kafka-python==2.2.15
Mako==1.3.10
MarkupSafe==3.0.3
minio==7.2.18
psycopg2==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
